requirements.txt
# Core ML / Embeddings
FlagEmbedding>=1.2.5           # BGE-M3 embedding model
sentence-transformers>=2.7.0,<3.0.0  # Cross-encoder + fallback embedder
transformers>=4.40.0,<4.50.0   # Pin to avoid broken TensorFlow DLL on Windows
faiss-cpu>=1.7.4               # FAISS vector index (use faiss-gpu if CUDA available)
rank-bm25>=0.2.2               # BM25 sparse retrieval

# LLM
google-generativeai>=0.8.0     # Gemini API client

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Caching
diskcache>=5.6.3

# File watching
watchdog>=4.0.0

# CLI / UI
rich>=13.7.0
streamlit>=1.35.0

# Progress bars
tqdm>=4.66.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
